+++ tests/utils.py
@@ -5,6 +5,8 @@ from mets_builder.metadata import Metadata
 from mets_builder.digital_object import DigitalObject
 from mets_builder.structural_map import StructuralMapDiv
 
+from siptools_ng.file import File
+
 
 def find_metadata(
     metadata_container: Union[DigitalObject, StructuralMapDiv],
@@ -16,8 +18,13 @@ def find_metadata(
     :param metadata_type: Metadata type to check for.
     :returns: Metadata that corresponds to the metadata_type.
     """
-    metadata = [
-        data for data in metadata_container.metadata
+    if isinstance(metadata_container, File):
+        metadata = metadata_container.digital_object.metadata
+    elif isinstance(metadata_container, StructuralMapDiv):
+        metadata = metadata_container.metadata
+
+    matching_metadata = [
+        data for data in metadata
         if isinstance(data, metadata_type)
     ][0]
-    return metadata
+    return matching_metadata