+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - Create a structural map automatically from directory structure when creating SIPs
 - Rename ``SIPDigitalObject`` to ``File``
 - Rename ``File`` class attribute ``source_filepath`` to ``path``
+- Rename ``File`` class attribute ``sip_filepath`` to ``digital_object_path``
 
 0.1.0 - 2024-03-27
 ------------------